Brought back the Scorable trait and move the wrapped_score method there for code organization
